refactor(lib): extract copyDirRecursive to shared file-utils.js (#duplication)
